Create user_landing module and move code to separate dir
