+++ handlers/go.mod
@@ -1,0 +1,3 @@
+module SocialMediaSite/handlers
+
+go 1.13
